C.O. Project (ETL+BI)/pyproject.toml
[project]
name = "container-offices-etl"
version = "0.2.0"
description = "ETL+BI pipeline for Container Offices data with history & forecasting"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.0.0",
    "polars>=0.19.0",
    "duckdb>=0.9.0",
    "pyarrow>=13.0.0",
    "openpyxl>=3.1.0",
    "python-dotenv>=1.0.0",
    "pandera>=0.17.0",
    "dbt-core>=1.6.0",
    "dbt-duckdb>=1.6.0",
    "ruff>=0.1.0",
    "black>=23.0.0",
    "pytest>=7.4.0",
    "statsmodels>=0.14.0",
    "pmdarima>=2.0.0",
    "scikit-learn>=1.3.0",
    "joblib>=1.3.0",
    "plotly>=5.17.0",
    "dash>=2.14.0",
]

[tool.setuptools]
packages = ["etl", "dashboard", "predict", "tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
